+++ NEWS.md
@@ -5,7 +5,7 @@
 ## New features
 
 * Including fitted and residual values as data frame in run_mod output. (#101)
-* Added `plot_predicted_curve()`  (#68)
+* Added  `plot_predicted_curve()` with support for faceting by multiple IDs (#68)
 * Replacing old data object with new run_mod output (#102)
 * Adding class assignment to run_mod output (#76)
 * Making prep_priors modifiable (#78)